+++ dbms/src/Flash/Coprocessor/GenSchemaAndColumn.cpp
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#include <Flash/Coprocessor/GenSchema.h>
+#include <Flash/Coprocessor/GenSchemaAndColumn.h>
 #include <Storages/Transaction/TypeMapping.h>
 
 namespace DB
